% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loi_normale.R
\name{loinormale_function}
\alias{loinormale_function}
\title{loi normale}
\usage{
loinormale_function(succes, moyenne, ecart_type)
}
\arguments{
\item{succes}{a number}

\item{moyenne}{a double}

\item{ecart_type}{a double}
}
\value{
a double
}
\description{
loi normale
}
