% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loi_probabilite.R
\name{loiprobabilite_function}
\alias{loiprobabilite_function}
\title{loi de probabilite}
\usage{
loiprobabilite_function(succes, epreuve, prob)
}
\arguments{
\item{succes}{a number}

\item{epreuve}{a number}

\item{prob}{a double}
}
\value{
a double
}
\description{
loi de probabilite
}
