% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_repartition.R
\name{repartition_function}
\alias{repartition_function}
\title{fontion de repartition}
\usage{
repartition_function(succes, epreuve, prob)
}
\arguments{
\item{succes}{a number}

\item{epreuve}{a number}

\item{prob}{a double}
}
\value{
a double
}
\description{
fontion de repartition
}
\examples{

# repartition_function <- function(succes,epreuve,prob){
dbinom(succes,epreuve,prob)
}
succes=5
epreuve=10
prob=0.3

}
