% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moy_ecart.R
\name{simulate_lln}
\alias{simulate_lln}
\title{Tcalcul moyenne et ecart type}
\usage{
simulate_lln(n, mu, sigma)
}
\arguments{
\item{n}{a number}

\item{mu}{a number}

\item{sigma}{a number}
}
\value{
a number
}
\description{
Tcalcul moyenne et ecart type
}
